Spectral Deconvolution
Personal Project that I accomplished. The code takes in the spectral data in line 134 (file location and the name) in the form of 2D array with X-axis values and Y-axis values for Raman spectra; the given data then becomes deconvoluted which is the technique widely used in signal and data processing. The code is independently written by me and my friend without external sources and the deconvoluted data is compared with nowadays software packages in the market; our code can produce similar or even more accurate results compraed to the nowadays tools on the market